fix issues with framework and configuration manager and also clean project
