--- JuanMartin.Api.Client.sln
@@ -1,30 +1,0 @@
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29926.136
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "JuanMartin.Models", "..\JuanMartin.Models\JuanMartin.Models\JuanMartin.Models.csproj", "{451CB413-E70B-4F70-93E9-76100453A956}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "JuanMartin.RestApiClient", "JuanMartin.RestClient\JuanMartin.RestApiClient.csproj", "{56CF3846-6D5B-4409-8DF8-A204F073D964}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{451CB413-E70B-4F70-93E9-76100453A956}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{451CB413-E70B-4F70-93E9-76100453A956}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{451CB413-E70B-4F70-93E9-76100453A956}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{451CB413-E70B-4F70-93E9-76100453A956}.Release|Any CPU.Build.0 = Release|Any CPU
-		{56CF3846-6D5B-4409-8DF8-A204F073D964}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{56CF3846-6D5B-4409-8DF8-A204F073D964}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{56CF3846-6D5B-4409-8DF8-A204F073D964}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{56CF3846-6D5B-4409-8DF8-A204F073D964}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {9EE9C97F-C2C0-4CCF-87A6-966B8997589E}
-	EndGlobalSection
-EndGlobal
