+++ JuanMartin.Api.sln
@@ -1,18 +1,15 @@
+﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
-VisualStudioVersion = 16.0.29926.136
+VisualStudioVersion = 16.0.31129.286
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Utilities", "..\JuanMartin.Utilities\JuanMartin.Utilities\Utilities.csproj", "{7E98C079-4E82-4A3F-8BD1-0BB0EFED02BF}"
-	ProjectSection(ProjectDependencies) = postProject
-		{451CB413-E70B-4F70-93E9-76100453A956} = {451CB413-E70B-4F70-93E9-76100453A956}
-	EndProjectSection
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "JuanMartin.Api", "JuanMartin.Api\JuanMartin.Api.csproj", "{3D830ACE-5A2E-4B45-94B6-778323586047}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "JuanMartin.Api", "JuanMartin.Api\JuanMartin.Api.csproj", "{5510308F-F4E2-48D4-B106-118EB0ED132A}"
-	ProjectSection(ProjectDependencies) = postProject
-		{451CB413-E70B-4F70-93E9-76100453A956} = {451CB413-E70B-4F70-93E9-76100453A956}
-	EndProjectSection
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "JuanMartin.Models", "..\JuanMartin.Models\JuanMartin.Models\JuanMartin.Models.csproj", "{6B59DE41-40F5-41C0-98CD-3393D463C757}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "JuanMartin.Models", "..\JuanMartin.Models\JuanMartin.Models\JuanMartin.Models.csproj", "{451CB413-E70B-4F70-93E9-76100453A956}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "JuanMartin.Utilities", "..\JuanMartin.Utilities\JuanMartin.Utilities\JuanMartin.Utilities.csproj", "{DDCBAA7C-5961-491F-8861-32003E400304}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "JuanMartin.Kernel", "..\JuanMartin.Kernel\JuanMartin.Kernel\JuanMartin.Kernel.csproj", "{4515E3A1-F6B3-4AF5-AE75-BB9E214EC349}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -20,23 +17,27 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7E98C079-4E82-4A3F-8BD1-0BB0EFED02BF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7E98C079-4E82-4A3F-8BD1-0BB0EFED02BF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7E98C079-4E82-4A3F-8BD1-0BB0EFED02BF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7E98C079-4E82-4A3F-8BD1-0BB0EFED02BF}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5510308F-F4E2-48D4-B106-118EB0ED132A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5510308F-F4E2-48D4-B106-118EB0ED132A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5510308F-F4E2-48D4-B106-118EB0ED132A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5510308F-F4E2-48D4-B106-118EB0ED132A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{451CB413-E70B-4F70-93E9-76100453A956}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{451CB413-E70B-4F70-93E9-76100453A956}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{451CB413-E70B-4F70-93E9-76100453A956}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{451CB413-E70B-4F70-93E9-76100453A956}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3D830ACE-5A2E-4B45-94B6-778323586047}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3D830ACE-5A2E-4B45-94B6-778323586047}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3D830ACE-5A2E-4B45-94B6-778323586047}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3D830ACE-5A2E-4B45-94B6-778323586047}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6B59DE41-40F5-41C0-98CD-3393D463C757}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6B59DE41-40F5-41C0-98CD-3393D463C757}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6B59DE41-40F5-41C0-98CD-3393D463C757}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6B59DE41-40F5-41C0-98CD-3393D463C757}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DDCBAA7C-5961-491F-8861-32003E400304}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DDCBAA7C-5961-491F-8861-32003E400304}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DDCBAA7C-5961-491F-8861-32003E400304}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DDCBAA7C-5961-491F-8861-32003E400304}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4515E3A1-F6B3-4AF5-AE75-BB9E214EC349}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4515E3A1-F6B3-4AF5-AE75-BB9E214EC349}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4515E3A1-F6B3-4AF5-AE75-BB9E214EC349}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4515E3A1-F6B3-4AF5-AE75-BB9E214EC349}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {9EE9C97F-C2C0-4CCF-87A6-966B8997589E}
+		SolutionGuid = {821C92AA-E5A1-4670-9DB9-89206D7F2957}
 	EndGlobalSection
 EndGlobal
