--- JuanMartin.Api/Properties/PublishProfiles/CustomProfile.pubxml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-This file is used by the publish/package process of your Web project. You can customize the behavior of this process
-by editing this MSBuild file. In order to learn more about this please visit https://go.microsoft.com/fwlink/?LinkID=208121. 
--->
-<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <WebPublishMethod>MSDeploy</WebPublishMethod>
-    <LastUsedBuildConfiguration>Release</LastUsedBuildConfiguration>
-    <LastUsedPlatform>Any CPU</LastUsedPlatform>
-    <SiteUrlToLaunchAfterPublish>http://localhost/JuanMartin.Api</SiteUrlToLaunchAfterPublish>
-    <LaunchSiteAfterPublish>True</LaunchSiteAfterPublish>
-    <ExcludeApp_Data>True</ExcludeApp_Data>
-    <MSDeployServiceURL>localhost</MSDeployServiceURL>
-    <DeployIisAppPath>Default Web Site/JuanMartin.Api</DeployIisAppPath>
-    <RemoteSitePhysicalPath />
-    <SkipExtraFilesOnServer>True</SkipExtraFilesOnServer>
-    <MSDeployPublishMethod>InProc</MSDeployPublishMethod>
-    <EnableMSDeployBackup>False</EnableMSDeployBackup>
-    <UserName />
-    <_SavePWD>False</_SavePWD>
-  </PropertyGroup>
-</Project>
