--- JuanMartin.Api/index.html
@@ -1,10 +1,0 @@
-﻿<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="utf-8" />
-    <title>JuanMartin Euler Solutions</title>
-</head>
-<body>
-    <h1><a href="https://projecteuler.net/about">Euler Project</a> solutions</h1>
-</body>
-</html>
