+++ JuanMartin.Api/packages.config
@@ -1,28 +1,53 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Microsoft.ApplicationInsights" version="2.13.1" targetFramework="net472" />
-  <package id="Microsoft.ApplicationInsights.Agent.Intercept" version="2.4.0" targetFramework="net472" />
-  <package id="Microsoft.ApplicationInsights.DependencyCollector" version="2.13.1" targetFramework="net472" />
-  <package id="Microsoft.ApplicationInsights.JavaScript" version="0.22.9-build00167" targetFramework="net452" />
-  <package id="Microsoft.ApplicationInsights.PerfCounterCollector" version="2.13.1" targetFramework="net472" />
-  <package id="Microsoft.ApplicationInsights.Web" version="2.13.1" targetFramework="net472" />
-  <package id="Microsoft.ApplicationInsights.WindowsServer" version="2.13.1" targetFramework="net472" />
-  <package id="Microsoft.ApplicationInsights.WindowsServer.TelemetryChannel" version="2.13.1" targetFramework="net472" />
-  <package id="Microsoft.AspNet.Mvc" version="5.2.7" targetFramework="net452" />
-  <package id="Microsoft.AspNet.Razor" version="3.2.7" targetFramework="net452" />
-  <package id="Microsoft.AspNet.TelemetryCorrelation" version="1.0.7" targetFramework="net472" />
-  <package id="Microsoft.AspNet.WebApi" version="5.2.7" targetFramework="net472" />
-  <package id="Microsoft.AspNet.WebApi.Client" version="5.2.7" targetFramework="net472" />
-  <package id="Microsoft.AspNet.WebApi.Core" version="5.2.7" targetFramework="net472" />
-  <package id="Microsoft.AspNet.WebApi.WebHost" version="5.2.7" targetFramework="net472" />
-  <package id="Microsoft.AspNet.WebPages" version="3.2.7" targetFramework="net452" />
-  <package id="Microsoft.CodeDom.Providers.DotNetCompilerPlatform" version="2.0.1" targetFramework="net472" />
-  <package id="Microsoft.Net.Compilers" version="3.5.0" targetFramework="net472" developmentDependency="true" />
-  <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" targetFramework="net452" />
-  <package id="Newtonsoft.Json" version="12.0.3" targetFramework="net472" />
+  <package id="Microsoft.AspNet.Mvc" version="5.2.7" targetFramework="net48" />
+  <package id="Microsoft.AspNet.Razor" version="3.2.7" targetFramework="net48" />
+  <package id="Microsoft.AspNet.TelemetryCorrelation" version="1.0.8" targetFramework="net48" />
+  <package id="Microsoft.AspNet.WebApi" version="5.2.7" targetFramework="net48" />
+  <package id="Microsoft.AspNet.WebApi.Client" version="5.2.7" targetFramework="net48" />
+  <package id="Microsoft.AspNet.WebApi.Core" version="5.2.7" targetFramework="net48" />
+  <package id="Microsoft.AspNet.WebApi.SelfHost" version="5.2.2" targetFramework="net48" />
+  <package id="Microsoft.AspNet.WebApi.WebHost" version="5.2.7" targetFramework="net48" />
+  <package id="Microsoft.AspNet.WebPages" version="3.2.7" targetFramework="net48" />
+  <package id="Microsoft.AspNetCore.Authentication.Abstractions" version="2.2.0" targetFramework="net48" />
+  <package id="Microsoft.AspNetCore.Authentication.Core" version="2.2.0" targetFramework="net48" />
+  <package id="Microsoft.AspNetCore.Authorization" version="2.2.0" targetFramework="net48" />
+  <package id="Microsoft.AspNetCore.Authorization.Policy" version="2.2.0" targetFramework="net48" />
+  <package id="Microsoft.AspNetCore.Hosting.Abstractions" version="2.2.0" targetFramework="net48" />
+  <package id="Microsoft.AspNetCore.Hosting.Server.Abstractions" version="2.2.0" targetFramework="net48" />
+  <package id="Microsoft.AspNetCore.Http" version="2.2.0" targetFramework="net48" />
+  <package id="Microsoft.AspNetCore.Http.Abstractions" version="2.2.0" targetFramework="net48" />
+  <package id="Microsoft.AspNetCore.Http.Extensions" version="2.2.0" targetFramework="net48" />
+  <package id="Microsoft.AspNetCore.Http.Features" version="2.2.0" targetFramework="net48" />
+  <package id="Microsoft.AspNetCore.Mvc.Abstractions" version="2.2.0" targetFramework="net48" />
+  <package id="Microsoft.AspNetCore.Mvc.Core" version="2.2.5" targetFramework="net48" />
+  <package id="Microsoft.AspNetCore.ResponseCaching.Abstractions" version="2.2.0" targetFramework="net48" />
+  <package id="Microsoft.AspNetCore.Routing" version="2.2.0" targetFramework="net48" />
+  <package id="Microsoft.AspNetCore.Routing.Abstractions" version="2.2.0" targetFramework="net48" />
+  <package id="Microsoft.AspNetCore.WebUtilities" version="2.2.0" targetFramework="net48" />
+  <package id="Microsoft.CodeDom.Providers.DotNetCompilerPlatform" version="3.6.0" targetFramework="net48" />
+  <package id="Microsoft.DotNet.PlatformAbstractions" version="2.1.0" targetFramework="net48" />
+  <package id="Microsoft.Extensions.Configuration.Abstractions" version="2.2.0" targetFramework="net48" />
+  <package id="Microsoft.Extensions.DependencyInjection" version="2.2.0" targetFramework="net48" />
+  <package id="Microsoft.Extensions.DependencyInjection.Abstractions" version="2.2.0" targetFramework="net48" />
+  <package id="Microsoft.Extensions.DependencyModel" version="2.1.0" targetFramework="net48" />
+  <package id="Microsoft.Extensions.FileProviders.Abstractions" version="2.2.0" targetFramework="net48" />
+  <package id="Microsoft.Extensions.Hosting.Abstractions" version="2.2.0" targetFramework="net48" />
+  <package id="Microsoft.Extensions.Logging.Abstractions" version="2.2.0" targetFramework="net48" />
+  <package id="Microsoft.Extensions.ObjectPool" version="2.2.0" targetFramework="net48" />
+  <package id="Microsoft.Extensions.Options" version="2.2.0" targetFramework="net48" />
+  <package id="Microsoft.Extensions.Primitives" version="2.2.0" targetFramework="net48" />
+  <package id="Microsoft.Net.Compilers" version="3.9.0" targetFramework="net48" developmentDependency="true" />
+  <package id="Microsoft.Net.Http.Headers" version="2.2.0" targetFramework="net48" />
+  <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" targetFramework="net48" />
+  <package id="Newtonsoft.Json" version="13.0.1" targetFramework="net48" />
   <package id="System.Buffers" version="4.5.1" targetFramework="net472" />
-  <package id="System.Diagnostics.DiagnosticSource" version="4.7.0" targetFramework="net472" />
+  <package id="System.ComponentModel.Annotations" version="4.5.0" targetFramework="net48" />
+  <package id="System.Diagnostics.DiagnosticSource" version="5.0.1" targetFramework="net48" />
   <package id="System.Memory" version="4.5.4" targetFramework="net472" />
   <package id="System.Numerics.Vectors" version="4.5.0" targetFramework="net472" />
-  <package id="System.Runtime.CompilerServices.Unsafe" version="4.7.1" targetFramework="net472" />
+  <package id="System.Runtime.CompilerServices.Unsafe" version="5.0.0" targetFramework="net48" />
+  <package id="System.Runtime.InteropServices.RuntimeInformation" version="4.0.0" targetFramework="net48" />
+  <package id="System.Text.Encodings.Web" version="4.5.0" targetFramework="net48" />
+  <package id="System.Threading.Tasks.Extensions" version="4.5.1" targetFramework="net48" />
 </packages>
