+++ JuanMartin.RestApiClient.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31129.286
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "JuanMartin.Models", "..\JuanMartin.Models\JuanMartin.Models\JuanMartin.Models.csproj", "{6B59DE41-40F5-41C0-98CD-3393D463C757}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "JuanMartin.RestApiClient", "JuanMartin.RestApiClient\JuanMartin.RestApiClient.csproj", "{C5BF1B64-5775-4F29-B273-03B3DA75173D}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{6B59DE41-40F5-41C0-98CD-3393D463C757}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6B59DE41-40F5-41C0-98CD-3393D463C757}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6B59DE41-40F5-41C0-98CD-3393D463C757}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6B59DE41-40F5-41C0-98CD-3393D463C757}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C5BF1B64-5775-4F29-B273-03B3DA75173D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C5BF1B64-5775-4F29-B273-03B3DA75173D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C5BF1B64-5775-4F29-B273-03B3DA75173D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C5BF1B64-5775-4F29-B273-03B3DA75173D}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {0C9F8D11-EE90-4052-A2D6-10E45A3028E1}
+	EndGlobalSection
+EndGlobal
