+++ JuanMartin.RestApiClient/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Newtonsoft.Json" version="13.0.1" targetFramework="net48" />
+  <package id="RestSharp" version="106.11.7" targetFramework="net48" />
+</packages>
